extract utility function aggregateIndex, numberCompare
