+++ src/model/CompositeNumberColumn.ts
@@ -5,8 +5,9 @@
 import {format} from 'd3';
 import Column, {IColumnDesc} from './Column';
 import CompositeColumn from './CompositeColumn';
-import NumberColumn, {INumberColumn, numberCompare} from './NumberColumn';
+import NumberColumn, {INumberColumn} from './NumberColumn';
 import {isMissingValue} from './missing';
+import {numberCompare} from './utils';
 
 export interface ICompositeNumberDesc extends IColumnDesc {
   /**
